% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_carte_utils.R
\name{extract_carte}
\alias{extract_carte}
\title{Extrait les valeurs de tous les rasters d'une liste à partir de coordonnées, pour une couche donnée}
\usage{
extract_carte(liste_place, carte, variable, couche)
}
\arguments{
\item{liste_place}{Table de 3 colonnes, une ligne par point à extraire.
\itemize{
 \item id_pe: identifiant du point
 \item latitude: latitude du point en degrés décimales (4326)
 \item longitude: longitude du point en degrés décimales (4326)
 }}

\item{carte}{Nom de la liste contenant les rasters}

\item{variable}{Vecteur contenant le nom des variables à extraire, par exemple: c("TMean","TotalPrecipitation")}

\item{couche}{Le nuémro de la couche à extraire de chaque raster, 1 si seulement une couche}
}
\value{
Table avec autant de colonnes que de rasters dans \code{carte}
}
\description{
\code{extract_carte} extrait les valeurs  de tous les rasters d'une liste à partir de coordonnées, pour une couche donnée
}
\details{
Pour chaque raster d'une liste, la fonction extrait les valeurs associées aux coordonnées des points. Les valeurs extraites de chaque raster sont regroupées dans une table, une colonne correspondant à un raster/couche
}
\examples{
\dontrun{
sol <- extract_carte(liste_place=liste_points, carte=cartes_sol, variable=c("cec","ph"), couche=1)
}
}
