% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_climat_an.R
\name{extract_climat_an}
\alias{extract_climat_an}
\title{Fonction pour extraire à partir de rasters et d'une liste de coordonnées, des variables de climat annuel et calculer la moyenne sur un nombre donnée d'années}
\usage{
extract_climat_an(file, variable, periode = 10)
}
\arguments{
\item{file}{Table de 4 colonnes, une ligne par point à extraire. La table peut contenir plus d'une ligne par id_pe, comme une liste d'arbres regroupés en placette. La fonction utilise les coordonnées des placettes.
\itemize{
 \item id_pe: identifiant du point
 \item latitude: latitude du point en degrés décimales (4326)
 \item longitude: longitude du point en degrés décimales (4326)
 \item an_mes: année correspondant à la fin de la période
 }}

\item{variable}{Vecteur contenant le nom des variables à extraire (sans l'année), par exemple: c("growingseasonprecipitation","growingseasontmean"). Variables disponibles:
\itemize{
 \item growingseasonprecipitation
 \item growingseasontmean
 }}

\item{periode}{Nombre d'années dans la période pour calculer la moyenne. Par exemple, si an_mes=2007, une période de 10 ans sera la moyenne de 1997 à 2006. Par défaut=10. Maximum 10.}
}
\value{
Table \code{file} avec les colonnes supplémentaires spécifiées dans \code{variable}
}
\description{
Extrait à partir de cartes de climat sous forme de raster et d'une liste de coordonnées, des variables de climat annuel et calcule la moyenne sur x années, pour les x années avant l'année spécifiée.
}
\details{
Il y a un raster par variable par année, de 1980 à 2022. Les rasters sont dans des fichiers tif dans le package.
Conditions: an_mes - periode >= 1980 et an_mes<=2023
}
\examples{
\dontrun{
climate_values <- extract_climat_an(file=fic_test, variable=c("growingseasonprecipitation","growingseasontmean"))
}
}
