% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_carte_utils.R
\name{extract_point}
\alias{extract_point}
\title{Extrait les valeurs d'un raster à partir de coordonnées}
\usage{
extract_point(fic_carte, element, couche, fic_points)
}
\arguments{
\item{fic_carte}{Nom de la liste contenant les rasters}

\item{element}{Numéro de l'élément de la liste correspodant au raster à utiliser}

\item{couche}{Numéro de la couche du raster. 1 si seulement une couche}

\item{fic_points}{Table sf contenant les coordonnées sous forme de geométrie, dans la projection appropriée, des points à extraire. Une ligne par point, 2 colonnes, lat et long.}
}
\value{
Table avec une colonne, le nom de la colonne est la concaténation du nom du raster et de la couche.
}
\description{
\code{extract_point} extrait les valeurs d'un raster à partir de coordonnées.
}
\examples{
\dontrun{
value <- extract_point(fic_carte=carte, element=1, couche=1, fic_points=liste_points)
}
}
