% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ExtractMap-package}
\alias{ExtractMap}
\alias{ExtractMap-package}
\title{ExtractMap: Extract values in maps from a set of coordinates}
\description{
Functions to extract values in maps from a set of coordinates. This package contains maps for iqs potentiels, cartes SIIGSOL, climat normales 30 ans, climat annuel.
}
\author{
\strong{Maintainer}: Isabelle Auger \email{isabelle.auger@mrnf.gouv.qc.ca}

}
\keyword{internal}
