% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fic_test}
\alias{fic_test}
\title{Fichier contenant des coordonnées}
\format{
## `fic_test`
A data frame with 29 lignes et 13 colonnes:
\describe{
  \item{id_pe}{Identifiant de la placette}
  \item{latitude}{Latitude de la placette}
  \item{longitude}{Latitude de la placette}
  \item{an_mes}{Année de mesure}
}
}
\usage{
fic_test
}
\description{
Deux placettes avec leur coordonnées et l'année de mesure
}
\examples{
fic_test
}
\keyword{datasets}
