% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_map_plot.R
\name{extract_map_plot}
\alias{extract_map_plot}
\title{Fonction pour effectuer des extractions à partir de raster et d'une liste de coordonnées}
\usage{
extract_map_plot(file, liste_raster, variable, profondeur = 1)
}
\arguments{
\item{file}{Table de 3 colonnes, une ligne par point à extraire. La table peut contenir plus d'une ligne par id_pe, comme une liste d'arbres regroupés en placette. La fonction utilise les coordonnées des placettes.
\itemize{
 \item id_pe: identifiant du point
 \item latitude: latitude du point en degrés décimales (4326)
 \item longitude: longitude du point en degrés décimales (4326)
 }}

\item{liste_raster}{Type de cartes:
\itemize{
  \item "cartes_iqs" : iqs pontentiels
  \item "cartes_sol" : propriétés de sol SIIGSOL
  \item "cartes_climat" : climats normales 30 ans
  }}

\item{variable}{Vecteur contenant le nom des variables à extraire, par exemple: c("tmean","totalprecipitation")
\itemize{
  \item cartes_iqs: iqs_pot_bop, iqs_pot_pex, iqs_pot_epb, iqs_pot_epn, iqs_pot_epr, iqs_pot_pig, iqs_pot_pib, iqs_pot_tho, iqs_pot_sab
  \item cartes_sol : cec, mat_org, ph, sable, limon, argile
  \item cartes_climat :
    \itemize{
  \item aridity
  \item consecutivedayswithoutfrost
  \item dayswithoutfrost
  \item degreeday
  \item firstfrostday
  \item growingseasonlength
  \item growingseasonprecipitation
  \item growingseasonradiation
  \item growingseasontmean
  \item julytmean
  \item lastfrostday
  \item pet
  \item snowfallproportion
  \item tmax
  \item tmean
  \item tmin
  \item totalprecipitation
  \item totalradiation
  \item totalsnowfall
  \item totalvpd
  \item utilprecipitation
  \item utilvpd
  }
  }}

\item{profondeur}{Profondeur pour les propriétés de sols, utilisé seulement si liste_raster="cartes_sol", 1 par défaut
\itemize{
 \item 1: profondeur 0-5 cm
 \item 2: pronfondeur 5-15 cm
}}
}
\value{
Table \code{file} avec les colonnes supplémentaires spécifiées dans \code{variable}
}
\description{
Extrait des valeurs à partir de raster et d'une liste de coordonnées
}
\details{
Il y a un raster par variable. Les rasters sont dans des fichiers tif dans le package.
}
\examples{
\dontrun{
soil_values <- extract_map_plot(file=fic_test, liste_raster="cartes_sol", variable=c("cec","mat_org"))
iqs_values <- extract_map_plot(file=fic_test, liste_raster="cartes_iqs", variable=c("iqs_pot_bop","iqs_pot_epn"))
climat_values <- extract_map_plot(file=fic_test, liste_raster="cartes_climat", variable=c("tmean","totalprecipitation"))
}
}
