% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_carte_main.R
\name{extract_map_plot}
\alias{extract_map_plot}
\title{Fonction principale pour effectuer des extractions à partir de raster et d'une liste de coordonnées}
\usage{
extract_map_plot(file, liste_raster, variable, couche)
}
\arguments{
\item{file}{Table de 3 colonnes, une ligne par point à extraire. La table peut contenir plus d'une ligne par id_pe, comme une liste d'arbres regroupés en placette. La fonction utilise les coordonnées des placettes.
\itemize{
 \item id_pe: identifiant du point
 \item latitude: latitude du point en degrés décimales (4326)
 \item longitude: longitude du point en degrés décimales (4326)
 }}

\item{liste_raster}{Nom de la liste contenant les rasters
\itemize{
  \item cartes_iqs : iqs pontentiels
  \item cartes_sol : propriétés de sol SIIGOLS
  \item cartes_climat : climats normales 30 ans
  }}

\item{variable}{Vecteur contenant le nom des variables à extraire, par exemple: c("tmean","totalprecipitation")
\itemize{
  \item cartes_iqs: iqs_pot_bop, iqs_pot_pex, iqs_pot_epb, iqs_pot_epn, iqs_pot_epr, iqs_pot_pig, iqs_pot_pib, iqs_pot_tho, iqs_pot_sab
  \item cartes_sol : cec, oc, ph, sand, silt, clay
  \item cartes_climat :
    \itemize{
  \item aridity
  \item consecutiveDayswithoutfrost
  \item dayswithoutfrost degreeday
  \item firstfrostday
  \item growingseasonlength
  \item growingseasonprecipitation
  \item growingseasonradiation
  \item growingseasontmean
  \item julytmean
  \item lastfrostday
  \item pet
  \item snowfallproportion
  \item tmax
  \item tmean
  \item tmin
  \item totalprecipitation
  \item totalradiation
  \item totalsnowfall
  \item totalvpd
  \item utilprecipitation
  \item utilvpd
  }
  }}

\item{couche}{Le nuémro de la couche à extraire de chaque raster
\itemize{
  \item Pour cartes_iqs, une seule couche
  \item Pour cartes_sol, 6 couches (1: 0-5cm, 2: 5-15cm, 3: 15-30cm, 4: 30-60cm, 5: 60-100cm, 6: 100-200cm)
  \item Pour cartes_climat, une seule couche
  }}
}
\value{
Table \code{file} avec les colonnes supplémentaires spécifiées dans \code{variable}
}
\description{
Extrait des valeurs à partir de raster et d'une liste de coordonnées
}
\examples{
\dontrun{
soil_values <- extract_map_plot(file=fic_test, liste_raster=cartes_sol, variable=c("cec","oc"), couche=1)
iqs_values <- extract_map_plot(file=fic_test, liste_raster=cartes_iqs, variable=c("iqs_pot_bop","iqs_pot_epn"), couche=1)
climat_values <- extract_map_plot(file=fic_test, liste_raster=cartes_climat, variable=c("tmean","totalprecipitation"), couche=1)
}
}
